Test out separate scss prop for custom pragma
