+++ packages/theme-ui/src/layout.js
@@ -26,7 +26,7 @@ export const Flex = styled(Box)({
 export const Layout = props =>
   jsx(Box, {
     ...props,
-    css: {
+    scss: {
       minHeight: '100vh',
       display: 'flex',
       flexDirection: 'column',
@@ -38,7 +38,7 @@ export const Header = props =>
   jsx(Box, {
     as: 'header',
     ...props,
-    css: {
+    scss: {
       display: 'flex',
       variant: 'styles.Header',
     }
@@ -47,7 +47,7 @@ export const Header = props =>
 export const Main = props =>
   jsx(Box, {
     ...props,
-    css: {
+    scss: {
       flex: '1 1 auto',
       variant: 'styles.Main',
     }
@@ -56,7 +56,7 @@ export const Main = props =>
 export const Container = props =>
   jsx(Box, {
     ...props,
-    css: {
+    scss: {
       width: '100%',
       minWidth: 0,
       maxWidth: 1024,
@@ -70,7 +70,7 @@ export const Footer = props =>
   jsx(Box, {
     as: 'footer',
     ...props,
-    css: {
+    scss: {
       display: 'flex',
       variant: 'styles.Footer',
     }
